feat: 회원가입 코스 선택 구현
