+++ src/views/Login/index.js
@@ -1,9 +1,11 @@
 import styled from 'styled-components';
 import SignIn from '../../components/SignForm/SignIn';
-import SignUp from '../../components/SignForm/SignUp';
 import { useState } from 'react';
+import SignUp from '../../components/SignForm/SignUp';
+
 const Login = () => {
   const [formType, setFormType] = useState('SIGNIN');
+
   return (
     <div>
       <SigninContainer>
